Move SMIRNOFF-specific things into its own submodule, subclassing things in system/collections.py
